Add APN values for Google Fi
Wear OS devices need to get provisioned with Fi specific APNs as part of setup. There is no Tycho app/PARIS on the wear devices so we can't rely on App based configuration. 3P OEMs rely on AOSP for such configuration

APNs that need to be added for legacy and new IMSI series: h2g2(default) and IMS

Change-Id: I462909ea634ed8cb1bdb8f1ecf582f98f6bac654
BUG: 276984485