Merge internet and MMS APN for H3G to fix MMS issue
The network interface of MMS connection returned by modem
is same as the default connection because the same apn field
is used for both MMS APN and internet APN.
But telephony will tear down it when duplicate network
interface is detected. So user can't send MMS.

For H3G, internet apn won't be route to iWLAN. So merge internet
and MMS APN for non-iWLAN case, and not define internet apn for
iWLAN case.

Test: adb shell content query --uri content://telephony/carriers --where numeric='23806'
Test: adb shell content query --uri content://telephony/carriers --where numeric='24002'
Bug: 353645779
Change-Id: I3afaf4a3bb6301356c65d35f183e7ec7bda165fe